backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.13
python-dotenv==1.0.1
mcp==1.12.4
openai==2.14.0
pykrx==1.0.51
pandas==2.2.3
requests==2.32.3
lxml==5.3.0
APScheduler==3.10.4
pytz==2024.2
setuptools==75.8.0
yfinance==0.2.43
